Initial Upload for running tests
This is where I will test my algorithms for rankings lists and what not. It's easier to create the program in Java so I'll make sure to organize the logic here. Unfortunately, UI is MUCH harder in Java so I am just going to make a web application but run my tests with Java.